change repository layout #6
resolves #6
