+++ .gitignore
@@ -3,3 +3,4 @@ node_modules
 .DS_Store
 *.beam
 *.dump
+_build/
