+++ .vscodeignore
@@ -3,6 +3,7 @@ typings/**
 out/test/**
 test/**
 src/**
+lib/**
 **/*.map
 .gitignore
 tsconfig.json
