+++ apps/erlangbridge/src/erlangbridge.app.src
@@ -1,0 +1,12 @@
+{application, erlangbridge,
+ [{description, ""},
+  {vsn, "0.1.0"},
+  {registered, []},
+  {modules, []},
+  {applications, [
+                 kernel
+                 ,stdlib
+                 ]},
+  %{mod, {erlangbridge_app, []}},
+  {env, []}
+ ]}.
