+++ lib/ErlangShell.ts
