+++ lib/vscodeAdapter.ts
