+++ rebar.config
@@ -1,0 +1,11 @@
+{require_otp_vsn, "18.*"}.
+{lib_dirs, []}.
+
+{deps, [
+      ]
+}.
+
+{erl_opts, [debug_info]}.
+{edoc_opts,[{todo,true}]}.
+
+{eunit_opts, [verbose, {report,{eunit_surefire,[{dir,"."}]}}]}.
