+++ rebar3.bat
@@ -1,0 +1,3 @@
+@echo off
+rem set DEBUG=1
+escript.exe rebar3 %*
